205: moved from LaunchUiUtils to UiUtils
apart from something specific of launch ui

Task-Url: http://github.com/pitest/pitclipse/issues/205
