+++ bundles/org.pitest.pitclipse.launch.ui/src/org/pitest/pitclipse/launch/ui/LaunchShortcut.java
@@ -28,7 +28,7 @@ import org.eclipse.debug.core.ILaunchConfiguration;
 import org.eclipse.jdt.core.IJavaElement;
 import org.eclipse.jdt.core.ITypeRoot;
 import org.eclipse.ui.IEditorInput;
-import org.pitest.pitclipse.launch.ui.utils.PitclipseLaunchUiUtils;
+import org.pitest.pitclipse.ui.utils.PitclipseUiUtils;
 
 import com.google.common.collect.ImmutableList;
 
@@ -41,7 +41,7 @@ final class LaunchShortcut {
 
     static Function<ITypeRoot, Optional<IResource>> getCorrespondingResource() {
         return t -> 
-            Optional.ofNullable(PitclipseLaunchUiUtils.executeSafelyOrElse(
+            Optional.ofNullable(PitclipseUiUtils.executeSafelyOrElse(
                 t::getCorrespondingResource, null));
     }
 
@@ -49,7 +49,7 @@ final class LaunchShortcut {
         if (o instanceof IJavaElement) {
             return Optional.of((IJavaElement) o);
         }
-        return Optional.ofNullable(PitclipseLaunchUiUtils.tryToAdapt(o, IJavaElement.class));
+        return Optional.ofNullable(PitclipseUiUtils.tryToAdapt(o, IJavaElement.class));
     }
 
     static ImmutableList<ILaunchConfiguration> emptyLaunchConfiguration() {
