+++ bundles/org.pitest.pitclipse.launch.ui/src/org/pitest/pitclipse/launch/ui/utils/PitclipseLaunchUiUtils.java
@@ -16,8 +16,8 @@
 package org.pitest.pitclipse.launch.ui.utils;
 
 import org.eclipse.core.runtime.CoreException;
-import org.eclipse.core.runtime.IAdaptable;
 import org.pitest.pitclipse.launch.ui.PitMutatorsTab;
+import org.pitest.pitclipse.ui.utils.PitclipseUiUtils.RunnableWithCoreException;
 
 /**
  * A few utilities for the UI.
@@ -29,26 +29,6 @@ public class PitclipseLaunchUiUtils {
 
     private PitclipseLaunchUiUtils() {
         // Only static utility methods
-    }
-
-    @FunctionalInterface
-    public static interface RunnableWithCoreException {
-        void run() throws CoreException;
-    }
-
-    @FunctionalInterface
-    public static interface ProviderWithCoreException<T> {
-        T get() throws CoreException;
-    }
-
-    @FunctionalInterface
-    public static interface PredicateWithCoreException {
-        boolean test() throws CoreException;
-    }
-
-    @FunctionalInterface
-    public static interface RunnableWithCoreExceptionInterruptable {
-        void run() throws CoreException, InterruptedException;
     }
 
     /**
@@ -66,62 +46,4 @@ public class PitclipseLaunchUiUtils {
         }
     }
 
-    /**
-     * Executes the passed predicate and in case of {@link CoreException} returns
-     * false.
-     * 
-     * @param predicate
-     */
-    public static boolean executeSafely(PredicateWithCoreException predicate) {
-        try {
-            return predicate.test();
-        } catch (CoreException ce) {
-            return false;
-        }
-    }
-
-    /**
-     * Returns the result of the passed provider and in case of {@link CoreException} returns
-     * the orElse argument.
-     * 
-     * @param predicate
-     */
-    public static <T> T executeSafelyOrElse(ProviderWithCoreException<T> provider, T orElse) {
-        try {
-            return provider.get();
-        } catch (CoreException ce) {
-            return orElse;
-        }
-    }
-
-    /**
-     * Executes the passed lambda and ignores exceptions.
-     * 
-     * @param runnable
-     */
-    public static void executeSafely(RunnableWithCoreExceptionInterruptable runnable) {
-        try {
-            runnable.run();
-        } catch (InterruptedException e) { // NOSONAR
-            // OK, silently move on
-        } catch (CoreException e) { // NOSONAR
-            // OK, silently move on
-        }
-    }
-
-    /**
-     * Try to adapt the object to the {@link IAdaptable} type, or returns null
-     * if it's not an {@link IAdaptable}.
-     * 
-     * @param <T>
-     * @param o
-     * @param type
-     * @return
-     */
-    public static <T extends IAdaptable> T tryToAdapt(Object o, Class<T> type) {
-        if (o instanceof IAdaptable) {
-            return ((IAdaptable) o).getAdapter(type);
-        }
-        return null;
-    }
 }
